Ajout des API pour la récupération des messages, des notifications, du client id et du client secret

+++ plugin/notification/Manager/NotificationManager.php
@@ -19,9 +19,11 @@ use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Icap\NotificationBundle\Library\ColorChooser;
 use JMS\DiExtraBundle\Annotation as DI;
 use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
+use RMS\PushNotificationsBundle\Message\AndroidMessage;
 
 /**
- * Class NotificationManager.
+ * Class NotificationManager
+ * @package Icap\NotificationBundle\Manager
  *
  * @DI\Service("icap.notification.manager")
  */
@@ -52,6 +54,37 @@ class NotificationManager
      */
     protected $notificationPluginConfigurationManager;
 
+    /**
+     * Constructor
+     * @DI\InjectParams({
+     *      "em" = @DI\Inject("doctrine.orm.entity_manager"),
+     *      "tokenStorage" = @DI\Inject("security.token_storage"),
+     *      "eventDispatcher" = @DI\Inject("event_dispatcher"),
+     *      "configHandler" = @DI\Inject("claroline.config.platform_config_handler"),
+     *      "notificationParametersManager" = @DI\Inject("icap.notification.manager.notification_user_parameters"),
+     *      "notificationPluginConfigurationManager" = @DI\Inject("icap.notification.manager.plugin_configuration")
+     * })
+     */
+    public function __construct(
+        EntityManager $em,
+        TokenStorageInterface $tokenStorage,
+        EventDispatcherInterface $eventDispatcher,
+        PlatformConfigurationHandler $configHandler,
+        NotificationUserParametersManager $notificationParametersManager,
+        NotificationPluginConfigurationManager $notificationPluginConfigurationManager
+    ) {
+        $this->em = $em;
+        $this->tokenStorage = $tokenStorage;
+        $this->eventDispatcher = $eventDispatcher;
+        $this->platformName = $configHandler->getParameter("name");
+        if ($this->platformName === null || empty($this->platformName)) {
+            $this->platformName = "Claroline";
+        }
+        $this->notificationParametersManager = $notificationParametersManager;
+        $this->notificationPluginConfigurationManager = $notificationPluginConfigurationManager;
+    }
+
+
     private function getLoggedUser()
     {
         $doer = null;
@@ -68,9 +101,7 @@ class NotificationManager
     private function getConfigurationAndPurge()
     {
         $config = $this->notificationPluginConfigurationManager->getConfigOrEmpty();
-        if ($config->getPurgeEnabled()) {
-            $this->purgeNotifications($config);
-        }
+        if($config->getPurgeEnabled())$this->purgeNotifications($config);
 
         return $config;
     }
@@ -95,7 +126,7 @@ class NotificationManager
         $iconKeys = $this->getNotificationRepository()->findAllDistinctIconKeys();
         $colorChooser = new ColorChooser();
         foreach ($iconKeys as $key) {
-            $colorChooser->getColorForName($key['iconKey']);
+            $colorChooser->getColorForName($key["iconKey"]);
         }
 
         return $colorChooser;
@@ -170,13 +201,13 @@ class NotificationManager
                 $notificationColor = $colorChooser->getColorForName($iconKey);
                 $notification->setIconColor($notificationColor);
             }
-            $eventName = 'create_notification_item_'.$notification->getActionKey();
+            $eventName = 'create_notification_item_' . $notification->getActionKey();
             $event = new NotificationCreateDelegateViewEvent($notificationView, $this->platformName);
 
-            /* @var EventDispatcher $eventDispatcher */
+            /** @var EventDispatcher $eventDispatcher */
             if ($this->eventDispatcher->hasListeners($eventName)) {
                 $event = $this->eventDispatcher->dispatch($eventName, $event);
-                $views[$notificationView->getId().''] = $event->getResponseContent();
+                $views[$notificationView->getId() . ''] = $event->getResponseContent();
             }
             if ($notificationView->getStatus() == false) {
                 array_push(
@@ -187,39 +218,11 @@ class NotificationManager
         }
         $this->markNotificationsAsViewed($unviewedNotificationIds);
 
-        return array('views' => $views, 'colors' => $colorChooser->getColorObjectArray());
+        return array("views" => $views, "colors" => $colorChooser->getColorObjectArray());
     }
 
-    /**
-     * Constructor.
-     *
-     * @DI\InjectParams({
-     *      "em" = @DI\Inject("doctrine.orm.entity_manager"),
-     *      "tokenStorage" = @DI\Inject("security.token_storage"),
-     *      "eventDispatcher" = @DI\Inject("event_dispatcher"),
-     *      "configHandler" = @DI\Inject("claroline.config.platform_config_handler"),
-     *      "notificationParametersManager" = @DI\Inject("icap.notification.manager.notification_user_parameters"),
-     *      "notificationPluginConfigurationManager" = @DI\Inject("icap.notification.manager.plugin_configuration")
-     * })
-     */
-    public function __construct(
-        EntityManager $em,
-        TokenStorageInterface $tokenStorage,
-        EventDispatcherInterface $eventDispatcher,
-        PlatformConfigurationHandler $configHandler,
-        NotificationUserParametersManager $notificationParametersManager,
-        NotificationPluginConfigurationManager $notificationPluginConfigurationManager
-    ) {
-        $this->em = $em;
-        $this->tokenStorage = $tokenStorage;
-        $this->eventDispatcher = $eventDispatcher;
-        $this->platformName = $configHandler->getParameter('name');
-        if ($this->platformName === null || empty($this->platformName)) {
-            $this->platformName = 'Claroline';
-        }
-        $this->notificationParametersManager = $notificationParametersManager;
-        $this->notificationPluginConfigurationManager = $notificationPluginConfigurationManager;
-    }
+
+
 
     /**
      * @return EntityManager
@@ -234,13 +237,13 @@ class NotificationManager
      */
     public function getPlatformName()
     {
-        return $this->platformName;
+       return $this->platformName;
     }
 
     /**
-     * Get Hash for a given object which must implement notifiable interface.
+     * Get Hash for a given object which must implement notifiable interface
      *
-     * @param int    $resourceId
+     * @param int $resourceId
      * @param string $resourceClass
      *
      * @return string The generated hash
@@ -257,7 +260,7 @@ class NotificationManager
     }
 
     /**
-     * @param int    $resourceId
+     * @param int $resourceId
      * @param string $resourceClass
      *
      * @return mixed
@@ -275,12 +278,12 @@ class NotificationManager
     }
 
     /**
-     * Create new Tag given its name.
+     * Create new Tag given its name
      *
-     * @param string      $actionKey
-     * @param string      $iconKey
-     * @param int|null    $resourceId
-     * @param array       $details
+     * @param string $actionKey
+     * @param string $iconKey
+     * @param integer|null $resourceId
+     * @param array $details
      * @param object|null $doer
      *
      * @internal param \Icap\NotificationBundle\Entity\NotifiableInterface $notifiable
@@ -306,11 +309,11 @@ class NotificationManager
 
         if (!isset($details['doer']) && !empty($doerId)) {
             $details['doer'] = array(
-                'id' => $doerId,
+                'id'        => $doerId,
                 'firstName' => $doer->getFirstName(),
-                'lastName' => $doer->getLastName(),
-                'avatar' => $doer->getPicture(),
-                'publicUrl' => $doer->getPublicUrl(),
+                'lastName'  => $doer->getLastName(),
+                'avatar'    => $doer->getPicture(),
+                'publicUrl' => $doer->getPublicUrl()
             );
         }
         $notification->setDetails($details);
@@ -323,11 +326,10 @@ class NotificationManager
     }
 
     /**
-     * Creates a notification viewer for every user in the list of people to be notified.
+     * Creates a notification viewer for every user in the list of people to be notified
      *
      * @param Notification $notification
      * @param $userIds
-     *
      * @internal param \Icap\NotificationBundle\Entity\NotifiableInterface $notifiable
      *
      * @return \Icap\NotificationBundle\Entity\Notification
@@ -342,7 +344,21 @@ class NotificationManager
                     $notificationViewer->setViewerId($userId);
                     $notificationViewer->setStatus(false);
 
+                    //push
+
+                    // $message = new AndroidMessage();
+                    // $message->setGCM(true);
+                    // $message->setMessage("coucou");
+                    // $message->setDeviceIdentifier('192.168.56.101:5555');
+                    //
+                    // $this->container->get('rms_push_notifications')->send($message);
+
+
+
+
                     $this->getEntityManager()->persist($notificationViewer);
+
+
                 }
             }
         }
@@ -352,10 +368,9 @@ class NotificationManager
     }
 
     /**
-     * Creates a notification and notifies the concerned users.
+     * Creates a notification and notifies the concerned users
      *
-     * @param NotifiableInterface $notifiable
-     *
+     * @param  NotifiableInterface $notifiable
      * @return Notification
      */
     public function createNotificationAndNotify(NotifiableInterface $notifiable)
@@ -384,14 +399,12 @@ class NotificationManager
     public function getDropdownNotifications($userId)
     {
         $config = $this->getConfigurationAndPurge();
-
         return $this->getUserNotificationsList($userId, 1, $config->getDropdownItems());
     }
 
     public function getPaginatedNotifications($userId, $page = 1, $category = null)
     {
         $config = $this->getConfigurationAndPurge();
-
         return $this->getUserNotificationsList($userId, $page, $config->getMaxPerPage(), false, null, $category);
     }
 
@@ -401,31 +414,21 @@ class NotificationManager
     }
 
     /**
-     * Retrieves the notifications list.
+     * Retrieves the notifications list
      *
-     * @param int  $userId
-     * @param int  $page
-     * @param int  $maxResult
-     * @param bool $isRss
-     *
+     * @param  int  $userId
+     * @param  int  $page
+     * @param  int  $maxResult
+     * @param  bool $isRss
      * @throws \Symfony\Component\HttpKernel\Exception\NotFoundHttpException
-     *
      * @return mixed
      */
     public function getUserNotificationsList($userId, $page = 1, $maxResult = -1, $isRss = false, $notificationParameters = null, $category = null)
     {
-        if ($notificationParameters == null) {
-            $notificationParameters = $this
-                ->notificationParametersManager
-                ->getParametersByUserId($userId);
-        }
-        $visibleTypes = $notificationParameters->getDisplayEnabledTypes();
-        if ($isRss) {
-            $visibleTypes = $notificationParameters->getRssEnabledTypes();
-        }
-        $query = $this
-            ->getNotificationViewerRepository()
-            ->findUserNotificationsQuery($userId, $visibleTypes, $category);
+
+        $query = $this->getUserNotifications($userId, $page, $maxResult, $isRss, $notificationParameters, false);
+
+
         $adapter = new DoctrineORMAdapter($query, false);
         $pager = new Pagerfanta($adapter);
         $pager->setMaxPerPage($maxResult);
@@ -439,10 +442,30 @@ class NotificationManager
         $notifications = $this->renderNotifications($pager->getCurrentPageResults());
 
         return array(
-            'pager' => $pager,
-            'notificationViews' => $notifications['views'],
-            'colors' => $notifications['colors'],
+            'pager'             => $pager,
+            'notificationViews' => $notifications["views"],
+            'colors'            => $notifications["colors"]
         );
+    }
+
+    public function getUserNotifications($userId, $page = 1, $maxResult = -1, $isRss = false, $notificationParameters = null, $executeQuery = true)
+    {
+        if ($notificationParameters == null) {
+            $notificationParameters = $this
+                ->notificationParametersManager
+                ->getParametersByUserId($userId);
+        }
+
+        if ($isRss) {
+            $visibleTypes = $notificationParameters->getRssEnabledTypes();
+        }
+        $visibleTypes = $notificationParameters->getDisplayEnabledTypes();
+
+        $query = $this
+            ->getNotificationViewerRepository()
+            ->findUserNotificationsQuery($userId, $visibleTypes);
+
+        return $executeQuery ? $query->getResult(): $query;
     }
 
     public function getUserNotificationsListRss($rssId)
@@ -451,7 +474,7 @@ class NotificationManager
         $notificationUserParameters = $this
             ->notificationParametersManager
             ->getParametersByRssId($rssId);
-        if ($notificationUserParameters === null) {
+        if($notificationUserParameters === null) {
             throw new NoResultException();
         }
 
@@ -465,8 +488,8 @@ class NotificationManager
     }
 
     /**
-     * @param int    $userId
-     * @param int    $resourceId
+     * @param int $userId
+     * @param int $resourceId
      * @param string $resourceClass
      *
      * @return null|object
@@ -476,7 +499,7 @@ class NotificationManager
         $followerResource = $this->getFollowerResourceRepository()->findOneBy(
             array(
                 'followerId' => $userId,
-                'hash' => $this->getHash($resourceId, $resourceClass),
+                'hash'       => $this->getHash($resourceId, $resourceClass)
             )
         );
 
@@ -485,13 +508,13 @@ class NotificationManager
 
     public function getTaggedUsersFromText($text)
     {
+
     }
 
     /**
      * @param $userId
      * @param $resourceId
      * @param $resourceClass
-     *
      * @return FollowerResource
      */
     public function followResource($userId, $resourceId, $resourceClass)
@@ -512,7 +535,6 @@ class NotificationManager
      * @param $userId
      * @param $resourceId
      * @param $resourceClass
-     *
      * @return mixed
      */
     public function unfollowResource($userId, $resourceId, $resourceClass)
@@ -538,8 +560,7 @@ class NotificationManager
     }
 
     /**
-     * @param null $viewerId
-     *
+     * @param  null $viewerId
      * @return int
      */
     public function countUnviewedNotifications($viewerId = null)
@@ -549,6 +570,6 @@ class NotificationManager
         }
         $notificationParameters = $this->notificationParametersManager->getParametersByUserId($viewerId);
 
-        return intval($this->getNotificationViewerRepository()->countUnviewedNotifications($viewerId, $notificationParameters->getDisplayEnabledTypes())['total']);
+        return intval($this->getNotificationViewerRepository()->countUnviewedNotifications($viewerId, $notificationParameters->getDisplayEnabledTypes())["total"]);
     }
 }
